requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.3
redis>=4.5.0
celery>=5.3.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
django-cors-headers>=4.6.0
djangorestframework-simplejwt>=3.15.2
django-filter>=24.3
python-dotenv>=1.0.1